test(mcp): move the WAV fixture out of a test file
`sampling-e2e.test.ts` imported `wav()` from `sampling.test.ts`, which made Bun
attribute all 20 of that file's tests to the importer: the e2e file reported 36
tests (its own 16 plus 20 pulled in) while `sampling.test.ts` reported none of
its own in any combined run. Every test still executed exactly once — the scoped
suite is 695 either way — but per-file totals were misleading. A plain helper
module keeps the attribution honest.
